requirements.txt
fastapi
uvicorn
line-bot-sdk
firebase-admin
pydantic
python-dotenv
requests